Use HTTPS URLs for submodules
To be able to checkout the repositories in an unauthenticated way (which
seems reasonable for a public project :), we need to use HTTPS URLs, as
they are the only ones GitHub support for unauthenticated access.

After this commit is applied you have to run `git submodule sync` for
git to update your fork URLs.